fix(embedding): make_labels returned an empty Labels instead of refusing
Review finding 8. An `embedding` model predicts appearance vectors, which have
no `sio.Labels` representation: `predict(make_labels=True)` and
`predict_to_file` bypassed the skeleton guard, found nothing to package, and
handed back an EMPTY `Labels` — indistinguishable from a model that predicted
nothing. Both now raise, naming `predict_embeddings_to_h5` and
`--embeddings_path`. The CLI already routed correctly; this is the Python API.

Also moves `minimal_embedding_model_dir` (and its config builder) from
`tests/export/conftest.py` into `tests/fixtures/model_ckpts.py`, which
`tests/conftest.py` star-imports, so the inference suite builds the same
loadable embedding model instead of growing a second copy of that config.
